Add support for logfmt (#830)
* Move hardcoded log format argument up to Logger

Instead of hardcoding the value in the extension, put it in the
Logger.log/5 function.

* Pass :format through Logger backend

Logger.log/5 now accepts a format, which defaults to :plaintext. To not
rely on that default, the Logger.Backend explicitly sets it.

* Add support for :logfmt

If the :format option is set to :logfmt, set the third argument in
Nig.log/5 to 1, indicating to the extension that these logs are in
logfmt format.