Create Finch span without root span
Even if there is no current active span, create a Finch root span.
This behaviour should be easier to understand and debug for users,
and should lead to request counts that better match their
expectations.

Use the `finch` namespace to avoid polluting the main performance
view with Finch root spans.