Implement Tesla instrumentation
Implement an instrumentation for the Tesla HTTP client library,
based on our existing Finch instrumentation.

If the Tesla.Middleware.PathParams middleware is used, a template is
used as the URL, and parameters are passed separately, to be
interpolated by this middleware. If the instrumentation detects that
the PathParams middleware is being used correctly, it will send the
template URL as part of the sanitised URL. Otherwise, it will only
send the scheme, host and port of the URL.

If the Tesla.Middleware.BaseUrl middleware is used, it will be used
to form the sanitised URL. Its path will be included in the sanitised
URL, as it is assumed to be static. If the PathParams middleware is
also being used, the template URL's path fragment will be appended to
the base URL. This allows for some "better than nothing" grouping for
users who use BaseUrl but not PathParams (e.g. `github.com/users`,
instead of just `github.com`)

Caveats
-------

This instrumentation relies on Tesla.Middleware.Telemetry being used
before any other middleware. The instrumentation benefits from seeing
what the user-intended request was, before the middleware modifies it.
Using other middlewares before Telemetry degrades its functionality.

If PathParams is used before Telemetry, the template will already have
been interpolated, resulting in a high-cardinality URL. To prevent
this, the order of the middlewares is taken into account to determine
whether PathParams is being used correctly and whether to include the
path fragment of the URL as part of the URL.

If BaseUrl is used before Telemetry, it will have no effect, as the
URL will have already been prefixed with the base URL. The base URL
will be ignored.

If MethodOverride is used before Telemetry, the method shown will be
the actual HTTP method sent over the wire (GET or POST) instead of
the semantically intended request method, which is sent in the
`X-HTTP-Method-Override` header.

Using Tesla.Middleware.Telemetry before any other middlewares prevents
these functionality degradations.