Treat Tesla/Finch exception events as stop
An exception event in Tesla or Finch is not commonly raised. If raised,
it would represent an exception happening inside Tesla or Finch. We have
not been able to manually test these events.

The instrumentations for Tesla and Finch will only create events in
already instrumented contexts. Therefore, catching these hypothetical
events and adding them as AppSignal errors in the Tesla or Finch
instrumentation is redundant -- the instrumented context can report the
error, if it actually results in a failure for the trace.

Remove the code that deals with Finch and Tesla exception events,
instead treating them as stop events.