Add elixir 1.15 to CI (#851)
* Revert "Don't run tests on Elixir main on CI

This reverts commit 63313291bc473f76064bfb36cc98fad8291548dc.
Closes #815.

* Use rebar as dependency manager for ssl_verify_fun

The ssl_verify_fun library doesn't depend on ssl, which is required
since Elixir 1.15. A patch[1] has been merged, but it hasn't been
released yet.

Until a new version of ssl_verify_fun is released, use rebar as the
dependenvy manager for ssl_verify_fun.

[1]: https://github.com/deadtrickster/ssl_verify_fun.erl/pull/27

* Add Elixir 1.15.0-rc.1 to CI versions

Add 1.15.0 to .semaphore/versions.rb

* Require Appsignal.Agent in diagnose_test

To resolve the compilation error in file
test/mix/tasks/appsignal_diagnose_test.exs on Elixir 1.15, explicitly
require Appsignal.Agent before referring to it.

* Add decimal to dependencies

Updating to Elixir 1.15 produces the following errors when running
dialyzer:

    Error loading decimal, dependency list may be incomplete.
     {~c"no such file or directory", ~c"decimal.app"}
    Error loading decimal, dependency list may be incomplete.
     {~c"no such file or directory", ~c"decimal.app"}

* Use Logger.warning/1 instead of warn/1

Logger.warn/1 is deprecated in favor of Logger.warning/1. This patch
replaces all occurrences of the former with the latter.

* Add backwards compatible logger

Since Logger.warn is now deprecated, add logging functions to
AppSignal.Utils that call Logger.warn on Elixir versions before 1.11,
and Logger.warning for newer versions.

This patch can be removed when dropping support for Elixir 1.10.

* Ignore dialyzer warnings introduced in Elixir 1.15

The new dialyzer warnings show an issue in the return types of the Nif's
diagnose function that can never be reached. Ignore these errors for now
in order to fix the build and ship the changes needed to run CI on
Elixir 1.15.