Remove calls to Config.active?/0 from Span
Currently, the Config.active?/0 function is called every time
Span.set_name/2 or Span.do_add_error/4 functions are called. Because
calling either of these functions on a nil is a noop, and nils are
passed around instead of spans when the package is not active, these can
be safely removed.

This reduces the amount of ets lookups by 2n when calling the
instrument/3 function with a name and a category.