Add Elixir 1.15 and Elixir main to CI (#861)
* Revert "Don't run tests on Elixir main on CI

This reverts commit 63313291bc473f76064bfb36cc98fad8291548dc.
Closes #815.

* Use rebar as dependency manager for ssl_verify_fun

The ssl_verify_fun library doesn't depend on ssl, which is required
since Elixir 1.15. A patch[1] has been merged, but it hasn't been
released yet.

Until a new version of ssl_verify_fun is released, use rebar as the
dependenvy manager for ssl_verify_fun.

[1]: https://github.com/deadtrickster/ssl_verify_fun.erl/pull/27

* Add Elixir 1.15.0 to CI versions

Add 1.15.0 to .semaphore/versions.rb

* Run linters and dialyzer on Elixir 1.15

Instead of running formatters and dialyzer on Elixir main, which is
prone to change, run these on the latest proper release.

* Format argument_cleaner_test to use ~c sigil

Since Elixir 1.15, the formatter uses the ~c sigil over single quotes
for char lists.

* Remove ssl_verify_fun override

A new version of ssl_verify_fun was released (1.1.7), which doesn't
require rebar to be used as a dependency manager anymore.

* Update diagnose test submodule

Update the diagnose tests submodule to update ssl_verify_fun to a new
version that works on Elixir 1.15 and up.