Remove unneeded _APPSIGNAL_DIAGNOSE env vars
The agent and extension no longer needs these env vars. It works without
it, because it now tell the agent to run the diagnose command by using
the `appsignal-agent diagnose` subcommand (as called by the extension).

I came across this when taking a brief look at the related dialyzer
issues reported in #853.

Update dialyzer warning because the stacktrace changed.

[skip changeset] because it's a small internal refactor.