Optional custom function on_create_span/2
With this function users can add their own function that will be
executed right after create_span is called. This allows users to inject
data into spans with traces of other tracing providers.

Example
```ex
  defmodule MyApp.Appsignal do
    def custom_on_create_fun(span, parent) do
      Appsignal.Span.set_sample_data(span, "custom_data", %{"foo": "bar"})
    end
  end
```
Then it can be added to the config with
```ex
  config :appsignal, custom_on_create_fun: &MyApp.Appsignal.custom_on_create_fun/2
```

After opening the PR  [Optional fun argument for attach](https://github.com/appsignal/appsignal-elixir-phoenix/pull/80)
in here @unflxw mentioned adding the option to this library and
executing it on the `create_span/3`.