Support encoding keyword lists
Add a test case for a previously failing scenario encoding keyword
lists.

Implement encoding for this by checking if a list is a keyword list,
then casting each item from a Tuple to a List, so the item is not
interpreted as a thing we can set on a Map.

If it encounters a keyword list's Tuple, it assumed the parent would
be a Map, and try to set it on the non-existent map. Then the Rust
Extension's data encoder would silently drop the value.