Remove parent as custom function parameter
Because:
Attributes can only be added to the root span and the parent isn't
guaranteed to be the root. When running the custom function in
on_create_span the root can be found with `Tracer.root_span/1`

The custom_custom_on_fun/1 no long needs to return a `Span.t() | nil'
and the outcome won't be assigned/used in the `on_create_span/1`.