Turn error backend off by default
The error backend is a relic of the past, which caught otherwise
uncaught errors before we had proper instrumentation through
library-specific instrumentation and :telemetry.

Also, the error backend produced confusing results, as the errors it
reports are void of context, making them hard to understand at best.

This patch turns the error backend off by default if it's unconfigured,
in preparation of removing the backend in a future version.