Drop non-500 errors in add_error/3 (#930)
Plug sets an attribute named plug_status on each of its raised
exceptions, which appsignal-plug uses to filter non-500 errors out. This
patch moves that filtering to the add_error/3 function, allowing the
integration to properly filter exceptions originating in plug, even if
they are not caught by the Plug integration.