Implement `Transmitter.transmit` method
Implement a `transmit` method, similar to that in the Ruby
integration, that sends a POST request with the API key, environment,
application name and hostname as query parameters.

Use it where requests with these parameters are sent, that is, for
the push API key validation request, and for the diagnose report
request.

Fix a bug where `config[:environment]` (a non-existent configuration
key) was accessed instead of `config[:env]`.