Assert if failing probe has run or not
This test I can't really tell how it asserts that the probe was not run
or that it not crashes the process, as it errors silently.

Add another probe that is run successfully and check if it has run more
than once. That means the other probe must have also run and raised an
error.