Add set_name_if_nil helper
Add helper to not set the name on a span if it's already set. We can use
this to only set the span name (and the root span's (action) name), to
not overwrite custom (action) names set by the app.

Add the name field to the span struct to avoid having to update the
extension and agent to handle this logic.

Part of #869