Add set_name_if_nil helper
Add helper to not set the name on a span if it's already set. We can use
this to only set the span name (and the root span's (action) name), to
not overwrite custom (action) names set by the app.

This change requires an update to the extension.
See PR https://github.com/appsignal/appsignal-agent/pull/1158

Part of #869