Remove error handling for IO.binwrite
Currently, the write_config_file/1 function in the installer checks
the return value of a call to IO.binwrite/2. However, that function
only ever returns :ok or raises an error when it fails to write.

This patch removes the error handling, assuming the write works. Since
the error handling is removed, the write will crash the installer if
it fails, which is the intended bahavior.