Remove deprecations for Utils functions
Instead of deprecating the no-longer-needed Appsignal.Utils.info/1,
Appsignal.Utils.warning/1, and Appsignal.Utils.compile_env/3, wait
until a new version of both Appsignal for Phoenix and Appsignal for
Plug are released.

That way, we won't run into a situation where a new version of
Appsignal for Elixir is released that produces deprecation warnings,
while the other libraries aren't out yet.

This allows for merging this patch without having to immediately do a
release.