Prefer the tracer span to the telemetry span
When processing an Ecto event, if both a tracer span and a telemetry
span are present, use the tracer span as the parent.

This should fix an issue where, when an Ecto transaction takes place
inside an Ecto parallel preload, when the transaction's commit or
rollback event attempts to close both its own span and the
overarching transaction span (which would be the tracer parent, within
the parallel preload's sub-process) it instead closes the transaction
span's parent (which would be the telemetry parent, from the process
that spawned the parallel preload) leaving the transaction span
unclosed and potentially cutting the transaction short.