Implement check-in scheduler
Implement a check-in event scheduler, which schedules check-in events
to be transmitted in a separate Elixir process, with a minimum wait
period of ten seconds between requests, and a wait period of a tenth
of a second before the first request, referred to as "debounce"
periods.

This is a relatively minor improvement for the existing cron
check-ins, but it is a requirement for the heartbeat check-ins, both
to avoid slowing down customers' applications with blocking requests,
and to avoid misuse of the feature from spamming our servers.

The scheduler also acts as a deduplicator, removing "similar enough"
check-in events -- again, not particularly interesting for cron
check-ins, but a requirement to minimise damage when heartbeat
check-ins are misused.

Use the previously implemented support for NDJSON payloads in the
transmitter in order to send all the check-ins in a single request.

Similar to the probes, an Elixir process is started by the AppSignal
application supervisor. When the process receives an event to be
transmitted, it stores it and schedules a message, which will trigger
the transmit, to be sent to itself at a later time.

When this Elixir process receives a shutdown signal, it attempts to
transmit all scheduled events before it shuts down.