Use separate Hackney pool for transmitter
This relates to the issue reported in #970, where the default Hackney
pool is drained of connections by the check-ins sent by the
transmitter. This does not fix the issue, but ensures that it is
a separate pool, used only by the AppSignal transmitter, that is
drained of connections, rather than the default one, which may be
used by the customer's code.