Close spans on error (#979)
When our `Appsignal.instrument` function encounters an error when the
function argument raises an error, the span it created would not be
closed.

Wrap the function call in a `try` block, and move the close logic to an
`after` block to ensure that the span is always closed.

I don't know if `instrument_root` is actually used anywhere. We don't
call it anywhere and we haven't documented it, but I'm updating it just
in case to match the `instrument` helper function.

Related internal thread:
https://appsignal.slack.com/archives/C7XHXQ7N0/p1739893354569399?thread_ts=1739870246.417719&cid=C7XHXQ7N0