Switch Hackney to Finch
Transition the HTTP client usage from Hackney to Finch to optimize
requests performance.

Two versions of transmissions are required now, as Finch needs to be run
under the supervision tree. One version is a standalone transmission
that starts and kills the process as soon as the requests is made, and
the other one uses the pool that's created when the integration starts.

Co-authored-by: Jeff Kreeftmeijer <jeff@kreeft.me>