Close the current span in instrument/2
In some situations, like in Phoenix apps, when the render/2 call is
done inside an instrument/2 block, the calls to Tracer.close_span/1
arrive out of order.

Previously, this would fail because the instrument/2 function tried to
close a span that's already been closed, resulting in situations where
the root span wasn't closed and nothing was reported.

This patch makes sure the current_span/0 is closed at the end of the
block, making sure the root span is closed properly.