Fix Finch usage in check-in scheduler
Remove `Transmitter.transmit_and_close` method. This method existed
due to some scenarios where a Hackney connection would be kept open
for its body to be read, saturating the Hackney pool.

As such, we previously changed how the transmitter worked, so that
either you call `.transmit`, and you're responsible for reading the
body and closing the request, or you call `.transmit_and_close`,
which takes care of closing the request for you, but does not
provide access to the body.

Finch does not seem to have such issues, and as such there is no
need for this convenience method.

In the push API key validator, don't assume that the error returned
by Finch (which is passed through by `Transmitter.transmit`) is
a `Mint.TransportError` with a `:reason` key. Prevent potential
pattern-matching errors by accepting differently-shaped errors as
well.

In the fake transmitter, which is used by the scheduler tests, make
the response returned by the fake transmission be Finch-shaped, not
Hackney-shaped. Fix the implementation accordingly.