Add mix appsignal.check_install task
In a support conversation we discussed how adding a mix task that can
check if AppSignal's NIF and agent were successfully installed.

We don't fail the installation if these couldn't be installed because we
don't want to breaks applications and prevent them from starting,
potentially bringing down whole applications. We rather not report data
to AppSignal in this scenario.

For people who do want this behavior, they can run this task before they
start their application or as part of their CI.

https://app.intercom.com/a/inbox/yzor8gyw/inbox/shared/all/conversation/16410700412377#part_id=comment-16410700412377-21909590891