Add namespace via attributes
In issue #972 it was reported it was difficult to override the default
set namespace because our Absinthe instrumentation also sets the
namespace to 'graphql', which is called later than some plugs.
We have no 'only set namespace if not already set' logic, because the
namespace is always set upon root span creation, so it would be
difficult to determine when the namespace value was changed.

In the linked agent PR (private link) we added logic for the Node.js and
Python OpenTelemetry integrations to set namespaces with the
`appsignal.namespace` and `appsignal.namespace_if_nil` attributes.
https://github.com/appsignal/appsignal-agent/pull/1274

This implementation also works for the Elixir integration, because they
all use the Span API.

If the `appsignal.namespace` attribute is set, it will use the value as
the namespace, overwriting the (default) namespace.
If only the `appsignal.namespace_if_nil` attribute is set, it will
use the value as the namespace, overwriting the default namespace.
If both are set, it will use the `appsignal.namespace` attribute value.

Since this uses the internal hoisting mechanism, it means the namespace
doesn't need to be set on the root span, but any span in the trace will
work. Not sure if we want to update the docs on this, or just keep
recommending to use the `Span.set_namespace` helper on the root span.

The Absinthe instrumentation (and potentially others) need to be updated
to use the `set_namespace_if_nil` helper so that it is easier to
overwrite the namespace for our users.