Allow customization of Absinthe trace namespace
Previously, it was not possible to customize the namespace of an
Absinthe trace as detailed in issue #972.
This was because the plug that set the namespace was run before our
Absinthe instrumentation, which would then overwrite the namespace
again.

We now use the new `Span.set_namespace_if_nil` helper in our Absinthe
instrumentation, so that it only set the namespace from the
instrumentation if it's not already set by the app (using the
`Span.set_namespace` helper).

Closes #972