Fix compilation on Elixir main (#1010)
It looks like the `do end` block no longer works for `catch_exit`.

The only thing it seems to do is catch an exit if one occurs, not assert
if it happens. If we remove it the tests fails on the errors we assert
on the spans that bubble up from the GenServer.

Directly pass in the GenServer call, like in the docs example:

```elixir
assert catch_exit(exit(1)) == 1
```

https://hexdocs.pm/ex_unit/ExUnit.Assertions.html#catch_exit/1